modify :: first
