+++ src/main/java/org/example/nova/NovaApplication.java
@@ -1,4 +1,4 @@
-package org.example.nova;
+package nova;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,8 +6,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class NovaApplication {
 
-    public static void main(String[] args) {
-        SpringApplication.run(NovaApplication.class, args);
-    }
+	public static void main(String[] args) {
+		SpringApplication.run(NovaApplication.class, args);
+	}
 
 }
