+++ src/main/java/org/example/nova/auth/info/OAuth2UserInfo.java
@@ -1,0 +1,8 @@
+package org.example.nova.auth.info;
+
+public interface OAuth2UserInfo {
+    String getProvider();
+    String getProviderId();
+    String getEmail();
+    String getName();
+}
