+++ src/main/java/org/example/nova/auth/presentation/LoginRequestDto.java
@@ -1,0 +1,14 @@
+package org.example.nova.auth.presentation;
+
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+public class LoginRequestDto {
+    private String loginId;
+    private String password;
+}
