--- src/main/java/org/example/nova/domain/auth/details/CustormOAuth2UserDetails.java
@@ -1,79 +1,0 @@
-package org.example.nova.domain.auth.details;
-
-import org.example.nova.domain.user.entity.User;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.oauth2.core.user.OAuth2User;
-
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Map;
-
-public class CustormOAuth2UserDetails implements UserDetails, OAuth2User {
-    private User user;
-    private Map<String, Object> attributes;
-
-    public void CustomUserDetails(User user, Map<String, Object> attributes) {
-
-        this.user = user;
-        this.attributes = attributes;
-    }
-
-    public CustormOAuth2UserDetails(User user, Map<String, Object> attributes) {
-        this.user = user;
-    }
-
-
-    @Override
-    public Map<String, Object> getAttributes() {
-        return attributes;
-    }
-
-    @Override
-    public String getName() {
-        return null;
-    }
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        Collection<GrantedAuthority> collection = new ArrayList<>();
-        collection.add(new GrantedAuthority() {
-            @Override
-            public String getAuthority() {
-                return user.getRole().name();
-            }
-        });
-
-        return collection;
-    }
-
-    @Override
-    public String getPassword() {
-        return user.getPassword();
-    }
-
-    @Override
-    public String getUsername() {
-        return user.getLoginId();
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-}
