--- src/main/java/org/example/nova/domain/auth/details/GoogleUserDetails.java
@@ -1,34 +1,0 @@
-package org.example.nova.domain.auth.details;
-
-import org.example.nova.domain.auth.info.OAuth2UserInfo;
-
-import java.util.Map;
-
-public class GoogleUserDetails implements OAuth2UserInfo {
-    private Map<String, Object> attributes;
-
-    public GoogleUserDetails(Map<String, Object> attributes) {
-        this.attributes = attributes;
-    }
-
-    @Override
-    public String getProvider() {
-        return "google";
-    }
-
-    @Override
-    public String getProviderId() {
-        return (String) attributes.get("sub");
-    }
-
-    @Override
-    public String getEmail() {
-        return (String) attributes.get("email");
-    }
-
-    @Override
-    public String getName() {
-        return (String) attributes.get("name");
-    }
-
-}
