--- src/main/java/org/example/nova/domain/auth/info/OAuth2UserInfo.java
@@ -1,8 +1,0 @@
-package org.example.nova.domain.auth.info;
-
-public interface OAuth2UserInfo {
-    String getProvider();
-    String getProviderId();
-    String getEmail();
-    String getName();
-}
