--- src/main/java/org/example/nova/domain/auth/presentation/LoginRequestDto.java
@@ -1,14 +1,0 @@
-package org.example.nova.domain.auth.presentation;
-
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-public class LoginRequestDto {
-    private String loginId;
-    private String password;
-}
