--- src/main/java/org/example/nova/domain/user/entity/UserRole.java
@@ -1,5 +1,0 @@
-package org.example.nova.domain.user.entity;
-
-public enum UserRole {
-    USER, ADMIN
-}
