+++ src/main/java/org/example/nova/home/dto/MissionRequestDto.java
@@ -1,0 +1,10 @@
+package org.example.nova.home.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class MissionRequestDto {
+    private int major;
+}
