+++ src/main/java/org/example/nova/home/dto/MissionResponseDto.java
@@ -1,0 +1,14 @@
+package org.example.nova.home.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class MissionResponseDto {
+    private String[] mission;
+    private int[] today;
+    private int level;
+    private String name;
+
+}
