+++ src/main/java/org/example/nova/home/dto/MissionTodayRequestDto.java
@@ -1,0 +1,10 @@
+package org.example.nova.home.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class MissionTodayRequestDto {
+    private int[] today;
+}
