+++ src/main/java/org/example/nova/home/dto/ProfileLevelResponseDto.java
@@ -1,0 +1,11 @@
+package org.example.nova.home.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class ProfileLevelResponseDto {
+    private int num;
+    private int level;
+}
