+++ src/main/java/org/example/nova/home/repository/ProfileRepository.java
@@ -1,0 +1,4 @@
+package org.example.nova.home.repository;
+
+public interface ProfileRepository {
+}
