+++ src/main/java/org/example/nova/user/entity/UserRole.java
@@ -1,0 +1,5 @@
+package org.example.nova.user.entity;
+
+public enum UserRole {
+    USER, ADMIN
+}
