+++ src/main/resources/application.yml
@@ -1,4 +1,3 @@
-
 spring:
 
   security:
