+++ src/main/resources/templates/login.html
@@ -8,7 +8,7 @@
 <body>
 <div>
     <h1><a th:href="|/${loginType}|">[[${pageName}]]</a></h1> <hr/>
-        <a href="/oauth2/authorization/google">구글 로그인</a>
+    <a href="/oauth2/authorization/google">구글 로그인</a>
     </form>
 </div>
 </body>
