+++ src/test/java/nova/NovaApplicationTests.java
@@ -1,0 +1,13 @@
+package nova;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class NovaApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
