--- src/test/java/org/example/nova/NovaApplicationTests.java
@@ -1,13 +1,0 @@
-package org.example.nova;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class NovaApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
